+++ lib/feature/profile/screen/my_profile_and_preferences_page.dart
@@ -1283,7 +1283,8 @@ class _MyProfileAndPreferencesContentState
       AppOverlays.showBrownyDialog(
         context,
         title: context.wording.errorOccurred,
-        message: 'ข้อมูลติดต่อไม่ถูกต้อง กรุณาลองใหม่อีกครั้ง',
+        // ข้อมูลติดต่อไม่ถูกต้อง กรุณาลองใหม่อีกครั้ง
+        message: context.wording.invalidContactInfoError,
       );
     }
   }
@@ -1297,7 +1298,8 @@ class _MyProfileAndPreferencesContentState
       AppOverlays.showBrownyDialog(
         context,
         title: context.wording.errorOccurred,
-        message: 'ข้อมูลติดต่อไม่ถูกต้อง กรุณาลองใหม่อีกครั้ง',
+        // ข้อมูลติดต่อไม่ถูกต้อง กรุณาลองใหม่อีกครั้ง
+        message: context.wording.invalidContactInfoError,
       );
     }
   }